requirements.txt
# Frontend & API
streamlit
fastapi
uvicorn
python-multipart
requests

# Computer Vision & OCR
pytesseract
Pillow

# AI & RAG Core
langchain
langchain-community
langchain-huggingface
langchain-core
sentence-transformers
faiss-cpu

# Utilities
thefuzz
python-dotenv
numpy